feat: Update package.json and enhance categories and listings tests; add test cases and remove deprecated files for improved test coverage
