+++ server/db.ts
@@ -1,5 +1,4 @@
 import { Pool } from "pg";
-import { config } from "dotenv";
 
 // Database connection pool
 export const pool = new Pool({
